Build out new canvas with ant x6
