+++ package.json
@@ -19,7 +19,9 @@
     "trailingComma": "all"
   },
   "dependencies": {
+    "@antv/layout": "^0.1.31",
     "@antv/x6": "^1.30.0",
+    "@antv/x6-react-components": "^1.1.15",
     "@antv/x6-react-shape": "^1.6.0",
     "@apollo/client": "^3.0.0-beta.39",
     "@apollo/react-hooks": "^3.1.5",
@@ -52,6 +54,7 @@
     "html-react-parser": "^0.10.3",
     "lodash": "^4.17.21",
     "next": "12.0.2",
+    "next-global-css": "^1.2.0",
     "next-themes": "^0.0.15",
     "next-with-apollo": "^5.2.1",
     "nookies": "^2.0.8",
@@ -84,7 +87,7 @@
     "@types/react": "^17.0.0",
     "@typescript-eslint/eslint-plugin": "^5.10.2",
     "@typescript-eslint/parser": "^5.10.2",
-    "autoprefixer": "^9.7.0",
+    "autoprefixer": "^10.4.2",
     "babel-loader": "^8.0.6",
     "css-loader": "^6.6.0",
     "eslint": "^8.8.0",
@@ -95,7 +98,7 @@
     "local-ssl-proxy": "^1.3.0",
     "ngrok": "^4.3.0",
     "patch-package": "^6.2.2",
-    "postcss": "^8.3.0",
+    "postcss": "^8.4.6",
     "postcss-import": "^14.0.2",
     "postinstall-postinstall": "^2.1.0",
     "sass": "^1.49.7",
@@ -103,7 +106,7 @@
     "storybook-addon-apollo-client": "^4.0.11",
     "storybook-addon-next": "^1.4.1",
     "style-loader": "^3.3.1",
-    "tailwindcss": "^3.0.7",
+    "tailwindcss": "^3.0.22",
     "ts-loader": "^9.2.6",
     "typescript": "^4.5.5",
     "vercel": "^23.1.2",
